# swagger.yaml
openapi: 3.0.3
info:
  title: Del-Pick API
  description: API for Del-Pick, a food delivery service application enabling customers to order food, drivers to deliver orders, and store owners to manage their menus and orders.
  version: 1.0.0
  contact:
    name: API Support
    email: support@delpick.id
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://delpick.horas-code.my.id/api/v1
    description: Production Server
  - url: http://localhost:6100/api/v1
    description: Development Server
...
paths:
...
  /menu-items:
    get:
      tags:
        - Menu Items
      summary: Retrieve all menu items
      description: Retrieves a paginated list of menu items, optionally filtered by store ID.
      operationId: getAllMenuItems
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/storeIdQueryParam'
      responses:
        '200':
          description: Menu items retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItemListResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Menu Items
      summary: Create a new menu item
      description: Creates a new menu item for a store. Requires store owner privileges.
      operationId: createMenuItem
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMenuItemRequest'
      responses:
        '201':
          description: Menu item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 201
                message: Menu item created successfully
                data:
                  id: 1
                  storeId: 1
                  name: Pizza
                  price: 10.99
                  isAvailable: true
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Not the store owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /menu-items/{menuItemId}:
    get:
      tags:
        - Menu Items
      summary: Retrieve menu item by ID
      description: Retrieves detailed information about a specific menu item.
      operationId: getMenuItemById
      parameters:
        - $ref: '#/components/parameters/menuItemIdParam'
      responses:
        '200':
          description: Menu item retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                message: Menu item retrieved successfully
                data:
                  id: 1
                  storeId: 1
                  name: Pizza
                  price: 10.99
                  isAvailable: true
        '404':
          description: Menu item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Menu Items
      summary: Update menu item
      description: Updates a menu item's information. Requires store owner privileges.
      operationId: updateMenuItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/menuItemIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMenuItemRequest'
      responses:
        '200':
          description: Menu item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                message: Menu item updated successfully
                data:
                  id: 1
                  storeId: 1
                  name: Pizza Updated
                  price: 11.99
                  isAvailable: true
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Not the store owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Menu item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Menu Items
      summary: Delete menu item
      description: Deletes a menu item. Requires store owner privileges.
      operationId: deleteMenuItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/menuItemIdParam'
      responses:
        '200':
          description: Menu item deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                message: Menu item deleted successfully
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Not the store owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Menu item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /menu-items/store/{storeId}:
    get:
      tags:
        - Menu Items
      summary: Retrieve menu items by store ID
      description: Retrieves a paginated list of menu items for a specific store.
      operationId: getMenuItemsByStoreId
      parameters:
        - $ref: '#/components/parameters/storeIdParam'
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        '200':
          description: Menu items retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItemListResponse'
        '404':
          description: Store not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Order Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    pageParam:
      name: page
      in: query
      description: Page number for pagination
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    limitParam:
      name: limit
      in: query
      description: Number of items per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
    storeIdParam:
      name: storeId
      in: path
      description: ID of the store
      required: true
      schema:
        type: integer
        minimum: 1
    storeIdQueryParam:
      name: storeId
      in: query
      description: Filter by store ID
      required: false
      schema:
        type: integer
        minimum: 1
    menuItemIdParam:
      name: menuItemId
      in: path
      description: ID of the menu item
      required: true
      schema:
        type: integer
        minimum: 1
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 200
        message:
          type: string
          example: Operation successful
        data:
          type: object
          nullable: true
      required:
        - statusCode
        - message
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: An error occurred
        errors:
          type: string
          nullable: true
          example: Detailed error message
      required:
        - statusCode
        - message

    # Authentication Schemas
...
    MenuItem:
      type: object
      properties:
        id:
          type: integer
          example: 1
        storeId:
          type: integer
          example: 1
        name:
          type: string
          example: Pizza
        description:
          type: string
          example: Classic Margherita
          nullable: true
        price:
          type: number
          format: float
          example: 10.99
        imageUrl:
          type: string
          format: uri
          example: https://example.com/menu/pizza.jpg
          nullable: true
        category:
          type: string
          example: Main Course
          nullable: true
        isAvailable:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2025-04-29T09:16:52.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-04-29T09:16:52.000Z"
    MenuItemListResponse:
      type: object
      properties:
        totalItems:
          type: integer
          example: 30
        totalPages:
          type: integer
          example: 3
        currentPage:
          type: integer
          example: 1
        menuItems:
          type: array
          items:
            $ref: '#/components/schemas/MenuItem'
    CreateMenuItemRequest:
      type: object
      required:
        - storeId
        - name
        - price
      properties:
        storeId:
          type: integer
          example: 1
        name:
          type: string
          maxLength: 100
          example: Pizza
        description:
          type: string
          maxLength: 500
          example: Classic Margherita
          nullable: true
        price:
          type: number
          format: float
          example: 10.99
        image:
          type: string
          description: Base64 encoded image
          example: data:image/jpeg;base64,/9j/4AAQSkZJRgABAQ...
          nullable: true
        category:
          type: string
          maxLength: 50
          example: Main Course
          nullable: true
        isAvailable:
          type: boolean
          example: true
          default: true
    UpdateMenuItemRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          example: Pizza Updated
          nullable: true
        description:
          type: string
          maxLength: 500
          example: Updated Margherita
          nullable: true
        price:
          type: number
          format: float
          example: 11.99
          nullable: true
        image:
          type: string
          description: Base64 encoded image
          example: data:image/jpeg;base64,/9j/4AAQSkZJRgABAQ...
          nullable: true
        category:
          type: string
          maxLength: 50
          example: Main Course
          nullable: true
        isAvailable:
          type: boolean
          example: true
          nullable: true

    # Order Schemas
...